requirements.txt
adblockparser==0.7
annotated-types==0.5.0
arrow==1.2.3
asgiref==3.7.2
autocommand==2.2.2
beautifulsoup4==4.12.2
binaryornot==0.4.4
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
cheroot==10.0.0
CherryPy==18.8.0
cherrypy-cors==1.7.0
click==8.1.7
colorama==0.4.6
cookiecutter==1.7.3
crispy-tailwind==0.5.0
cryptography==3.4.8
distlib==0.3.7
Django==3.2.6
django-crispy-forms==2.0
django-environ==0.4.5
django-sb-admin==5.5.1.1
django-tailwind==2.2.0
djangorestframework==3.14.0
dnspython==2.4.2
et-xmlfile==1.1.0
ExifRead==2.3.2
filelock==3.12.2
future==0.18.3
httpagentparser==1.9.5
idna==3.4
inflect==7.0.0
ipaddr==2.2.0
ipwhois==1.1.0
jaraco.collections==4.3.0
jaraco.context==4.3.0
jaraco.functools==3.9.0
jaraco.text==3.11.1
Jinja2==3.1.2
jinja2-time==0.2.0
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
more-itertools==10.1.0
netaddr==0.8.0
networkx==2.6.3
openpyxl==3.1.2
phonenumbers==8.13.20
Pillow==9.5.0
platformdirs==3.9.1
portend==3.2.0
poyo==0.5.0
publicsuffixlist==0.9.4
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pygexf==0.2.2
pyOpenSSL==21.0.0
PyPDF2==1.28.6
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.22
python-slugify==8.0.1
python-whois==0.7.3
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
secure==0.3.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tempora==5.5.0
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.24.0
XlsxWriter==3.1.3
zc.lockfile==3.0.post1
django-cors-headers
